requirements.txt
pyautogui
time
PIL.Image
pygetwindow
imutils
numpy
cv2
